requirements.txt
google-genai>=0.5.0
gradio>=4.0.0
numpy>=1.24.0
pygame>=2.5.0
websockets>=12.0
phonemizer>=3.2.0